Lift web document-scroll into the layout; slim per-screen hook to canvas colour
The iphone-web-preview work made native document scroll a per-screen opt-in:
every content screen called useWebDocumentScroll(colour), which bundled three
concerns — enabling document scroll, resetting scroll on navigation, and painting
the canvas. Two of those are identical on every content route, so a new screen
that forgets the hook silently inherits Expo's nested-scroll trap (exactly the
landing-page bug just fixed).

Move the universal mechanism into app/_layout.web.tsx: it switches the body to
native document scroll for every content route and resets scroll on navigation,
while (auth) screens stay app-shell (Expo's body{overflow:hidden}). New content
screens now get correct scrolling for free.

Rename the hook to useWebCanvas — it now owns only the part that legitimately
differs per screen: the canvas colour (one screen even derives it from responsive
state, so it must stay screen-local). Update all 10 call sites.
