+++ app/_layout.web.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import { View, StyleSheet } from 'react-native';
-import { Stack, useRouter, useSegments } from 'expo-router';
+import { Stack, useRouter, useSegments, usePathname } from 'expo-router';
 import { useFonts } from 'expo-font';
 import { StatusBar } from 'expo-status-bar';
 import { Nunito_400Regular, Nunito_700Bold, Nunito_900Black } from '@expo-google-fonts/nunito';
@@ -17,12 +17,14 @@ function WebAuthGate() {
   const { user, loading } = useAuth();
   const segments = useSegments();
   const router = useRouter();
+  const pathname = usePathname();
   const [settled, setSettled] = useState(false);
+
+  const segs = segments as string[];
+  const inAuthGroup = segs[0] === '(auth)';
 
   useEffect(() => {
     if (loading) return;
-    const segs = segments as string[];
-    const inAuthGroup = segs[0] === '(auth)';
     const isRoot = segs.length === 0;
 
     if (user && (inAuthGroup || isRoot)) {
@@ -30,12 +32,33 @@ function WebAuthGate() {
     } else {
       setSettled(true);
     }
-  }, [user, loading, segments, router]);
+  }, [user, loading, segs, inAuthGroup, router]);
+
+  // Native document scroll is the default for every content route: content bleeds
+  // edge-to-edge under the iOS Safari toolbar and the toolbar can collapse (it
+  // only minimizes on *document* scroll, not on a nested RNW ScrollView). Auth
+  // screens stay app-shell — full-height forms that keep Expo's body{overflow:
+  // hidden}. Owning this here means a new content screen gets correct scrolling
+  // without remembering to opt in; screens only declare their canvas colour via
+  // useWebCanvas.
+  useEffect(() => {
+    if (typeof document === 'undefined') return;
+    if (inAuthGroup) {
+      document.body.style.removeProperty('overflow');
+    } else {
+      document.body.style.setProperty('overflow', 'visible', 'important');
+    }
+  }, [inAuthGroup]);
+
+  // Document scroll keeps the window offset across navigation (unlike a per-screen
+  // ScrollView that always mounts at the top), so reset to the top on every route
+  // change — otherwise a new screen can open part-scrolled.
+  useEffect(() => {
+    if (typeof window !== 'undefined') window.scrollTo(0, 0);
+  }, [pathname]);
 
   if (loading || !settled) return <LogoLoader />;
 
-  const segs = segments as string[];
-  const inAuthGroup = segs[0] === '(auth)';
   const isRoot = segs.length === 0;
   const showNav = !inAuthGroup && !isRoot;
   // Every top-level page lets its dark header bleed up under the floating nav for
