+++ app/index.tsx
@@ -1,10 +1,10 @@
 import { Platform } from 'react-native';
 import { Redirect } from 'expo-router';
 import LandingPage from '../src/components/landing/LandingPage.dom';
-import { useWebDocumentScroll } from '../src/hooks/useWebDocumentScroll';
+import { useWebCanvas } from '../src/hooks/useWebCanvas';
 
 export default function Index() {
-  useWebDocumentScroll('#0b1820');
+  useWebCanvas('#0b1820');
   if (Platform.OS !== 'web') return <Redirect href="/explore" />;
   return <LandingPage dom={{ scrollEnabled: false, matchContents: true }} />;
 }
